requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
websockets>=11.0
requests>=2.31.0
numpy>=1.24.0
pandas>=2.0.0
yfinance>=0.2.0
Pillow>=10.0.0
PyYAML>=6.0
scipy>=1.11.0
librosa>=0.10.0
modelscope>=1.20.0
optimum-intel[openvino]>=1.18.0
transformers>=4.40.0
